Adapt UI to type changes
