+++ src/app/types/extraTypes.ts
@@ -1,3 +1,5 @@
+import { ConditionType } from '@types';
+
 export enum FlightControlMetrics {
   ACTIVE_AGENT_COUNT_METRIC = 'flightctl_devicesimulator_active_agent_count',
   TOTAL_API_REQUESTS_METRIC = 'flightctl_devicesimulator_api_requests_total',
@@ -29,5 +31,17 @@ export interface MetricsQuery {
   metrics: FlightControlMetrics[];
   period: string;
 }
+
+// We define separate "ConditionType" for each resource type, given it's generic in the API
+export type FleetConditionType = 'Synced' | 'Syncing'; // TODO reference values from the API when they are defined
+export type RepositoryConditionType =
+  | ConditionType.RepositoryAccessible
+  | ConditionType.ResourceSyncResourceParsed
+  | ConditionType.ResourceSyncAccessible;
+
+export type EnrollmentRequestApprovalStatus = 'Approved' | 'Pending' | 'Denied' | 'Unknown';
+export type FleetUpdateStatus = 'Synced' | 'Syncing' | 'Unknown';
+export type RepositorySyncStatus = RepositoryConditionType | 'NotSynced' | 'NotParsed' | 'NotAccessible' | 'Unknown';
+export type ResourceStatusType = FleetUpdateStatus | RepositorySyncStatus | EnrollmentRequestApprovalStatus;
 
 export type FlightControlQuery = ApiQuery | MetricsQuery;
